requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2025.8.3
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.1
cycler==0.11.0
datasets==3.3.2
debugpy==1.8.11
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.17.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
huggingface_hub==0.29.2
idna==3.7
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
nest_asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.1
numpy==2.0.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==5.9.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
Pygments==2.19.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.5
SQLAlchemy==2.0.41
stack_data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2023.3
unicodedata2==15.1.0
update_checker==0.18.0
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
